+++ tests/phpunit/Integration/I18nJsonFileIntegrityTest.php
@@ -37,6 +37,10 @@ class I18nJsonFileIntegrityTest extends \PHPUnit\Framework\TestCase {
 		$provider = [];
 		$location = $GLOBALS['wgMessagesDirs']['SemanticInterlanguageLinks'];
 
+		if ( is_array( $location ) ) {
+			$location = $location[0];
+		}
+
 		$bulkFileProvider = UtilityFactory::getInstance()->newBulkFileProvider( $location );
 		$bulkFileProvider->searchByFileExtension( 'json' );